# Proof of Concept: Outlined RawTable