@Gankro 2019-04-27
Also when running `cargo test` locally on my mac I see:

```
...
test map::test_map::test_occupied_entry_key ... ok
test map::test_map::test_remove ... ok
test map::test_map::test_remove_entry ... ok
test map::test_map::test_raw_entry ... ok
test map::test_map::test_show ... ok
hashbrown-94db3bfa716f052b(5235,0x70000df2e000) malloc: can't allocate region
*** mach_vm_map(size=13835058055282167808) failed (error code=3)
hashbrown-94db3bfa716f052b(5235,0x70000df2e000) malloc: *** set a breakpoint in malloc_error_break to debug
test map::test_map::test_size_hint ... ok
test map::test_map::test_retain ... ok
test map::test_map::test_try_reserve ... ok
test map::test_map::test_vacant_entry_key ... ok
...
```

This is troubling, but all the test pass. Is there just a test that's checking we fail to allocate a massive thing?