@Gankro 2019-04-27
cargo bench results:

before:

```
test insert_erase_i32 ... bench:      27,526 ns/iter (+/- 1,773)
test insert_erase_i64 ... bench:      30,098 ns/iter (+/- 5,014)
test insert_i32       ... bench:      21,701 ns/iter (+/- 1,867)
test insert_i64       ... bench:      24,181 ns/iter (+/- 2,313)
test iter_i32         ... bench:       1,794 ns/iter (+/- 131)
test iter_i64         ... bench:       1,805 ns/iter (+/- 103)
test lookup_fail_i32  ... bench:       1,821 ns/iter (+/- 149)
test lookup_fail_i64  ... bench:       1,795 ns/iter (+/- 157)
test lookup_i32       ... bench:       2,713 ns/iter (+/- 292)
test lookup_i64       ... bench:       2,802 ns/iter (+/- 361)
```

after:

```
running 10 tests
test insert_erase_i32 ... bench:      24,184 ns/iter (+/- 6,220)
test insert_erase_i64 ... bench:      31,264 ns/iter (+/- 4,505)
test insert_i32       ... bench:      18,267 ns/iter (+/- 1,264)
test insert_i64       ... bench:      25,695 ns/iter (+/- 1,993)
test iter_i32         ... bench:       1,797 ns/iter (+/- 136)
test iter_i64         ... bench:       1,791 ns/iter (+/- 173)
test lookup_fail_i32  ... bench:       2,059 ns/iter (+/- 90)
test lookup_fail_i64  ... bench:       2,061 ns/iter (+/- 87)
test lookup_i32       ... bench:       3,035 ns/iter (+/- 117)
test lookup_i64       ... bench:       3,215 ns/iter (+/- 389)
```