@Amanieu 2019-04-27
I'm a bit skeptical, this feels like a somewhat roundabout way of simply boxing the `HashMap`. The only difference is that you have 2 allocations instead of 1.