@Gankro 2019-04-27
remove ctrl bench:

```
test insert_erase_i32 ... bench:      24,478 ns/iter (+/- 665)
test insert_erase_i64 ... bench:      31,363 ns/iter (+/- 4,813)
test insert_i32       ... bench:      18,021 ns/iter (+/- 1,253)
test insert_i64       ... bench:      25,321 ns/iter (+/- 1,040)
test iter_i32         ... bench:       1,818 ns/iter (+/- 82)
test iter_i64         ... bench:       1,822 ns/iter (+/- 109)
test lookup_fail_i32  ... bench:       2,125 ns/iter (+/- 79)
test lookup_fail_i64  ... bench:       1,824 ns/iter (+/- 125)
test lookup_i32       ... bench:       2,979 ns/iter (+/- 76)
test lookup_i64       ... bench:       3,194 ns/iter (+/- 178)
```