@Gankro 2019-04-27
Well it's significantly better than `Box<HashMap>` for several reasons:

* One allocation
* Doesn't allocate for HashMap::new()
* with ctrl removed: same memory footprint as not-boxing

fwiw Swift's Dictionary type has a similar outlined design, although they obviously have fairly different constraints what with reference-counted CoW collections and ObjC interop.

At very least, the implementation is here, so it would definitely be interesting for parties interested in shrinking the footprint of HashMap to grab this and profile/benchmark their workloads with it. At least locally on my laptop it does seem to result in some hurting on `cargo bench`, but this sort of change is one which probably is more significant in non-synthetic situations, where memory usage and cache pressure are more significant (keep in mind the empty singleton is shared, so performing ops on it is more likely to be in cache).