@Gankro 2019-04-27
reran benchmarks, honestly not *too* bad a difference between no-ctrl and master. Close enough that some messaging might plausibly get us to parity.

Disclaimer: my laptop is very prone to thermal throttling, so someone with a more stable system might want to try it.

```
old:
test insert_erase_i32 ... bench:      28,049 ns/iter (+/- 2,981)
test insert_erase_i64 ... bench:      29,961 ns/iter (+/- 4,659)
test insert_i32       ... bench:      22,098 ns/iter (+/- 856)
test insert_i64       ... bench:      24,722 ns/iter (+/- 4,997)
test iter_i32         ... bench:       1,810 ns/iter (+/- 167)
test iter_i64         ... bench:       1,813 ns/iter (+/- 72)
test lookup_fail_i32  ... bench:       1,835 ns/iter (+/- 85)
test lookup_fail_i64  ... bench:       1,796 ns/iter (+/- 101)
test lookup_i32       ... bench:       2,758 ns/iter (+/- 153)
test lookup_i64       ... bench:       2,813 ns/iter (+/- 111)


new (ctrl):
test insert_erase_i32 ... bench:      29,437 ns/iter (+/- 834)
test insert_erase_i64 ... bench:      25,578 ns/iter (+/- 3,472)
test insert_i32       ... bench:      23,261 ns/iter (+/- 2,539)
test insert_i64       ... bench:      19,313 ns/iter (+/- 973)
test iter_i32         ... bench:       1,804 ns/iter (+/- 43)
test iter_i64         ... bench:       1,785 ns/iter (+/- 130)
test lookup_fail_i32  ... bench:       2,055 ns/iter (+/- 81)
test lookup_fail_i64  ... bench:       2,044 ns/iter (+/- 275)
test lookup_i32       ... bench:       3,170 ns/iter (+/- 208)
test lookup_i64       ... bench:       3,127 ns/iter (+/- 261)

new (no ctrl):
test insert_erase_i32 ... bench:      23,930 ns/iter (+/- 882)
test insert_erase_i64 ... bench:      30,739 ns/iter (+/- 4,469)
test insert_i32       ... bench:      17,866 ns/iter (+/- 710)
test insert_i64       ... bench:      24,982 ns/iter (+/- 679)
test iter_i32         ... bench:       1,782 ns/iter (+/- 141)
test iter_i64         ... bench:       1,790 ns/iter (+/- 171)
test lookup_fail_i32  ... bench:       2,049 ns/iter (+/- 146)
test lookup_fail_i64  ... bench:       1,793 ns/iter (+/- 151)
test lookup_i32       ... bench:       2,871 ns/iter (+/- 215)
test lookup_i64       ... bench:       3,116 ns/iter (+/- 161)
```